completed maker component
